requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.0.8
django-braces==1.14.0
django-filter==2.2.0
django-oauth-toolkit==1.3.2
django-rest-framework-multi-slug-field==0.2.0
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
timedelta==0.0.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
